requirements.txt
Django==2.2
djangorestframework==3.10.2
psycopg2==2.8.6
pytz==2022.4
sqlparse==0.4.3
